development-erp/erp-fastapi/requirements.txt
fastapi==0.115.6
python-dotenv==1.0.1
SQLAlchemy==2.0.40
asyncmy==0.2.10
python-multipart==0.0.20
pydantic-settings==2.7.1
bcrypt==4.2.1
python-jose==3.3.0
uvicorn==0.34.0
factory_boy==3.3.1
pytest-mock==3.14.0
pytest-asyncio==0.25.2
cryptography==44.0.0
